+++ pkg/sql/exec_factory_util.go
@@ -13,8 +13,12 @@ package sql
 import (
 	"fmt"
 
+	"github.com/cockroachdb/cockroach/pkg/sql/execinfra"
+	"github.com/cockroachdb/cockroach/pkg/sql/opt/cat"
 	"github.com/cockroachdb/cockroach/pkg/sql/opt/exec"
 	"github.com/cockroachdb/cockroach/pkg/sql/rowexec"
+	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
+	"github.com/cockroachdb/cockroach/pkg/sql/sqlbase"
 	"github.com/cockroachdb/errors"
 )
 
@@ -79,4 +83,24 @@ func constructPlan(
 	}
 
 	return res, nil
+}
+
+// makeScanColumnsConfig builds a scanColumnsConfig struct by constructing a
+// list of descriptor IDs for columns in the given cols set. Columns are
+// identified by their ordinal position in the table schema.
+func makeScanColumnsConfig(table cat.Table, cols exec.TableColumnOrdinalSet) scanColumnsConfig {
+	// Set visibility=execinfra.ScanVisibilityPublicAndNotPublic, since all
+	// columns in the "cols" set should be projected, regardless of whether
+	// they're public or non-public. The caller decides which columns to
+	// include (or not include). Note that when wantedColumns is non-empty,
+	// the visibility flag will never trigger the addition of more columns.
+	colCfg := scanColumnsConfig{
+		wantedColumns: make([]tree.ColumnID, 0, cols.Len()),
+		visibility:    execinfra.ScanVisibilityPublicAndNotPublic,
+	}
+	for c, ok := cols.Next(0); ok; c, ok = cols.Next(c + 1) {
+		desc := table.Column(c).(*sqlbase.ColumnDescriptor)
+		colCfg.wantedColumns = append(colCfg.wantedColumns, tree.ColumnID(desc.ID))
+	}
+	return colCfg
 }
